rename constants, remove comments
