+++ src/constants/folio-to-innreach-locations.js
@@ -1,6 +1,6 @@
 export const FOLIO_TO_INN_REACH_LOCATIONS_ROUTE = 'folio-to-inn-reach-locations';
 
-export const FOLIO_TO_INN_REACH_LOCATIONS = {
+export const FOLIO_TO_INN_REACH_LOCATION_FIELDS = {
   TABULAR_LIST: 'tabularList',
   LEFT_COLUMN: 'leftColumn',
   CENTRAL_SERVER: 'centralServerId',
